enh: enabled pkg-config creation if template pc/pc.in files are present
Now pkg-config files are installed if a <PROJECT_NAME>.pc/.pc.in file
is found in the TOP_DIR.

I have additionally converted all hardcoded exec-commands into
wrapper variables that allows custom executables. Apparently
this is needed on MacOSx.

Now there is full control from the user (with the exception of
flags passed to the install command). I have nothing to test with.

Signed-off-by: Nick Papior <nickpapior@gmail.com>